Protein Ligand classes now inherit from Molecule class; more functionality for Protein class too.
